createGhaDevice now returns an Expected
Creation of the device iotself has been refactored to no longer happen in the construct and instead happens in a free function and only returns a device object on success
