+++ clove/components/core/graphics/include/Clove/Graphics/Gha.hpp
@@ -2,8 +2,10 @@
 
 #include "Clove/Graphics/GraphicsAPI.hpp"
 
+#include <Clove/Expected.hpp>
 #include <any>
 #include <memory>
+#include <stdexcept>
 
 namespace clove {
     class GhaDevice;
@@ -11,9 +13,9 @@ namespace clove {
 
 namespace clove {
     /**
-     * @brief Create a Graphics Device to communicate with the graphics hardware.
+     * @brief Create a GHA (Graphics Hardware Api) Device to communicate with the graphics hardware.
      * @param api What API to use.
      * @param nativeWindow The window to bind to. Using an empty std::any will create a headless device.
      */
-    std::unique_ptr<GhaDevice> createGraphicsDevice(GraphicsApi api, std::any nativeWindow);
+    Expected<std::unique_ptr<GhaDevice>, std::runtime_error> createGhaDevice(GraphicsApi api, std::any nativeWindow) noexcept;
 }
