+++ clove/components/core/graphics/include/Clove/Graphics/Metal/DeviceWrapper.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "Clove/Graphics/Metal/MetalView.hpp"
+
+#include <MetalKit/MetalKit.h>
+
+namespace clove {
+    struct DeviceWrapper {
+        id<MTLDevice> device{ nullptr };
+        MetalView *view{ nullptr };
+    };
+}
