+++ clove/components/core/graphics/source/Metal/MetalDevice.mm
@@ -2,33 +2,14 @@
 
 #include "Clove/Graphics/Metal/MetalFactory.hpp"
 #include "Clove/Graphics/Metal/MetalView.hpp"
+#include "Clove/Graphics/Metal/DeviceWrapper.hpp"
+
+#include <MetalKit/MetalKit.h>
 
 namespace clove {
-    struct DeviceWrapper {
-        id<MTLDevice> device{ nullptr };
-        MetalView *view{ nullptr };
-    };
-}
-
-namespace clove {
-    MetalDevice::MetalDevice(std::any nativeWindow) {
-        wrapper = std::make_unique<DeviceWrapper>();
-        
-        //Create device
-        wrapper->device = MTLCreateSystemDefaultDevice();
-        
-        //Create view
-        if(nativeWindow.has_value()) {
-            NSWindow *nsWindow{ std::any_cast<NSWindow *>(nativeWindow) };
-            wrapper->view = [[MetalView alloc] initWithFrame:[nsWindow frame]];
-            [wrapper->view setDevice:wrapper->device];
-        
-            [nsWindow setContentView:wrapper->view];
-        }else {
-            wrapper->view = nullptr;
-        }
-        
-        factory = std::make_unique<MetalFactory>(wrapper->device, wrapper->view);
+    MetalDevice::MetalDevice(std::unique_ptr<DeviceWrapper> deviceWrapper, std::unique_ptr<MetalFactory> factory) 
+        : wrapper{ std::move(deviceWrapper) }
+        , factory{ std::move(factory) }{
     }
     
     MetalDevice::MetalDevice(MetalDevice &&other) noexcept = default;
